crypto-intelligence-agent/requirements.txt
httpx>=0.27.0
pyarrow>=15.0.0
snowflake-connector-python>=3.10.0
